Update Sample App
