+++ .editorconfig
@@ -1,0 +1,7 @@
+[*.cs]
+
+# Organize usings
+dotnet_sort_system_directives_first = true
+
+# CS4014: Because this call is not awaited, execution of the current method continues before the call is completed
+dotnet_diagnostic.CS4014.severity = error
