+++ samples/CommunityToolkit.Maui.Markup.Sample.sln
@@ -10,9 +10,10 @@ EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{FB2C97E0-5C47-4A80-BE66-6DE982933BD0}"
 	ProjectSection(SolutionItems) = preProject
 		Directory.Build.props = Directory.Build.props
-		Directory.Build.targets = Directory.Build.targets
 		global.json = global.json
 	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CommunityToolkit.Maui.Markup.UnitTests", "..\src\CommunityToolkit.Maui.Markup.UnitTests\CommunityToolkit.Maui.Markup.UnitTests.csproj", "{F45A2C29-DDD2-49A8-B4D9-57150F80AD36}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -30,6 +31,10 @@ Global
 		{A4511CE3-FBBE-467E-AC49-522D3D308189}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A4511CE3-FBBE-467E-AC49-522D3D308189}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A4511CE3-FBBE-467E-AC49-522D3D308189}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F45A2C29-DDD2-49A8-B4D9-57150F80AD36}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F45A2C29-DDD2-49A8-B4D9-57150F80AD36}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F45A2C29-DDD2-49A8-B4D9-57150F80AD36}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F45A2C29-DDD2-49A8-B4D9-57150F80AD36}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
