+++ samples/CommunityToolkit.Maui.Markup.Sample/App.cs
@@ -3,23 +3,13 @@ using CommunityToolkit.Maui.Markup.Sample.Pages;
 using Microsoft.Maui;
 using Microsoft.Maui.Controls;
 
-namespace CommunityToolkit.Maui.Markup.Sample
+namespace CommunityToolkit.Maui.Markup.Sample;
+
+class App : Application
 {
-    class App : Application
+    public App(NewsPage newsPage) => MainPage = new NavigationPage(newsPage)
     {
-        readonly NewsPage _newsPage;
-
-        public App(NewsPage newsPage) => _newsPage = newsPage;
-
-        protected override Window CreateWindow(IActivationState activationState)
-        {
-            var navigationPage = new NavigationPage(_newsPage)
-            {
-                BarBackgroundColor = ColorConstants.NavigationBarBackgroundColor,
-                BarTextColor = ColorConstants.NavigationBarTextColor
-            };
-
-            return new Window(navigationPage);
-        }
-    }
+        BarBackgroundColor = ColorConstants.NavigationBarBackgroundColor,
+        BarTextColor = ColorConstants.NavigationBarTextColor
+    };
 }
