+++ samples/CommunityToolkit.Maui.Markup.Sample/Constants/ColorConstants.cs
@@ -1,16 +1,15 @@
 ﻿using Microsoft.Maui.Graphics;
 
-namespace CommunityToolkit.Maui.Markup.Sample.Constants
+namespace CommunityToolkit.Maui.Markup.Sample.Constants;
+
+static class ColorConstants
 {
-    static class ColorConstants
-    {
-        public static Color NavigationBarBackgroundColor { get; } = Color.FromHex("FF6601");
-        public static Color NavigationBarTextColor { get; } = Colors.Black;
+    public static Color NavigationBarBackgroundColor { get; } = Color.FromHex("FF6601");
+    public static Color NavigationBarTextColor { get; } = Colors.Black;
 
-        public static Color TextCellDetailColor { get; } = Color.FromHex("828282");
-        public static Color TextCellTextColor { get; } = Colors.Black;
+    public static Color TextCellDetailColor { get; } = Color.FromHex("828282");
+    public static Color TextCellTextColor { get; } = Colors.Black;
 
-        public static Color BrowserNavigationBarBackgroundColor { get; } = Color.FromHex("FFE6D5");
-        public static Color BrowserNavigationBarTextColor { get; } = Color.FromHex("3F3F3F");
-    }
+    public static Color BrowserNavigationBarBackgroundColor { get; } = Color.FromHex("FFE6D5");
+    public static Color BrowserNavigationBarTextColor { get; } = Color.FromHex("3F3F3F");
 }
