+++ samples/CommunityToolkit.Maui.Markup.Sample/Constants/EmojiConstants.cs
@@ -1,10 +1,9 @@
-﻿namespace CommunityToolkit.Maui.Markup.Sample.Constants
+﻿namespace CommunityToolkit.Maui.Markup.Sample.Constants;
+
+static class EmojiConstants
 {
-    static class EmojiConstants
-    {
-        public const string SadFaceEmoji = "☹️";
-        public const string NeutralFaceEmoji = "\U0001F610";
-        public const string HappyFaceEmoji = "\U0001F603";
-        public const string BlankFaceEmoji = "\U0001F636";
-    }
+    public const string SadFaceEmoji = "☹️";
+    public const string NeutralFaceEmoji = "\U0001F610";
+    public const string HappyFaceEmoji = "\U0001F603";
+    public const string BlankFaceEmoji = "\U0001F636";
 }
