+++ samples/CommunityToolkit.Maui.Markup.Sample/Pages/Base/BasePage.cs
@@ -1,16 +1,15 @@
 ﻿using CommunityToolkit.Maui.Markup.Sample.ViewModels.Base;
 using Microsoft.Maui.Controls;
 
-namespace CommunityToolkit.Maui.Markup.Sample.Pages.Base
-{
-    abstract class BaseContentPage<T> : ContentPage where T : BaseViewModel
-    {
-        protected BaseContentPage(T viewModel, string pageTitle)
-        {
-            BindingContext = ViewModel = viewModel;
-            Title = pageTitle;
-        }
+namespace CommunityToolkit.Maui.Markup.Sample.Pages.Base;
 
-        protected T ViewModel { get; }
+abstract class BaseContentPage<T> : ContentPage where T : BaseViewModel
+{
+    protected BaseContentPage(T viewModel, string pageTitle)
+    {
+        BindingContext = ViewModel = viewModel;
+        Title = pageTitle;
     }
+
+    protected T ViewModel { get; }
 }
