+++ samples/CommunityToolkit.Maui.Markup.Sample/Platforms/Android/MainActivity.cs
@@ -2,10 +2,9 @@
 using Android.Content.PM;
 using Microsoft.Maui;
 
-namespace CommunityToolkit.Maui.Markup.Sample
+namespace CommunityToolkit.Maui.Markup.Sample;
+
+[Activity(Theme = "@style/Maui.SplashTheme", MainLauncher = true, ConfigurationChanges = ConfigChanges.ScreenSize | ConfigChanges.Orientation | ConfigChanges.UiMode | ConfigChanges.ScreenLayout | ConfigChanges.SmallestScreenSize)]
+public class MainActivity : MauiAppCompatActivity
 {
-	[Activity(Theme = "@style/Maui.SplashTheme", MainLauncher = true, ConfigurationChanges = ConfigChanges.ScreenSize | ConfigChanges.Orientation | ConfigChanges.UiMode | ConfigChanges.ScreenLayout | ConfigChanges.SmallestScreenSize)]
-	public class MainActivity : MauiAppCompatActivity
-	{
-	}
 }
