+++ samples/CommunityToolkit.Maui.Markup.Sample/Platforms/MacCatalyst/Program.cs
@@ -1,9 +1,8 @@
 ﻿using UIKit;
 
-namespace CommunityToolkit.Maui.Markup.Sample
+namespace CommunityToolkit.Maui.Markup.Sample;
+
+public class Program
 {
-    public class Program
-    {
-        static void Main(string[] args) => UIApplication.Main(args, null, typeof(AppDelegate));
-    }
+    static void Main(string[] args) => UIApplication.Main(args, null, typeof(AppDelegate));
 }
