+++ samples/CommunityToolkit.Maui.Markup.Sample/Platforms/iOS/Info.plist
@@ -5,7 +5,7 @@
 	<key>LSRequiresIPhoneOS</key>
 	<true/>
 	<key>MinimumOSVersion</key>
-	<string>10.3.4</string>
+	<string>14.2.0</string>
 	<key>UIDeviceFamily</key>
 	<array>
 		<integer>1</integer>
